Bugfix postbootcommandfile argument execution
Moved the bootcommand class to inside the server, since the required TranslatedConfigView class initialize inside the server
Remove prebootcommandfile argument because it is impossible to achieve due to it require TranslatedConfigView dependency that run inside the server

# Conflicts:
#	nucleus/core/kernel/src/main/java/com/sun/enterprise/v3/server/AppServerStartup.java
#	nucleus/payara-modules/hazelcast-bootstrap/src/main/java/fish/payara/nucleus/hazelcast/HazelcastCore.java
