+++ nucleus/core/kernel/src/test/java/com/sun/enterprise/v3/bootstrap/BootCommandsTest.java
@@ -1,7 +1,7 @@
 /*
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * Copyright (c) [2016-2019] Payara Foundation and/or its affiliates. All rights reserved.
+ * Copyright (c) [2019-2022] Payara Foundation and/or its affiliates. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common Development
@@ -37,51 +37,42 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package fish.payara.boot.runtime;
+package com.sun.enterprise.v3.bootstrap;
 
-import org.glassfish.embeddable.CommandResult;
-import org.glassfish.embeddable.CommandResult.ExitStatus;
-import org.glassfish.embeddable.CommandRunner;
-
-import static java.util.logging.Level.INFO;
-import static java.util.logging.Level.WARNING;
-import java.util.logging.Logger;
+import java.io.IOException;
+import java.io.Reader;
+import java.io.StringReader;
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertThat;
+import org.junit.Test;
 
 /**
- * Simple class for holding an asadmin command representation for using during
- * pre and post boot
- * @author Steve Millidge
+ *
+ * @author Gaurav Gupta
  */
-public class BootCommand {
-    
-    private final String command;
-    private final String[] arguments;
+public class BootCommandsTest {
 
-    private static final Logger LOGGER = Logger.getLogger(BootCommand.class.getName());
-
-    public BootCommand(String command, String ... arguments) {
-        this.command = command;
-        this.arguments = arguments;
-    }
-    
-    public boolean execute(CommandRunner runner) {
-        boolean result = true;
-        CommandResult asadminResult = runner.run(command, arguments);
-        if (asadminResult.getExitStatus().equals(ExitStatus.FAILURE)) {
-            LOGGER.log(WARNING, String.format("Boot Command %s failed %s ", command, asadminResult.getOutput()), asadminResult.getFailureCause());
-            result = false;
-        } else {
-            LOGGER.log(INFO, String.format("Boot Command %s returned with result %s : %s", command, asadminResult.getExitStatus(), asadminResult.getOutput()));
+    @Test
+    public void parseCommand() throws IOException {
+        BootCommands bootCommands = new BootCommands();
+        String commandText = "create-custom-resource --restype java.lang.String -s v --name='custom-res' --description=\"results \\\"in\\\" error\" --property value=\"${ENV=ini_ws_uri}\" vfp/vfp-menu/ini.ws.uri";
+        try (Reader reader = new StringReader(commandText)){
+            bootCommands.parseCommandScript(reader);
         }
-        return result;
-    }
-
-    public String getCommand() {
-        return command;
-    }
-
-    public String[] getArguments() {
-        return arguments;
+        assertThat(bootCommands.getCommands().size(), is(1));
+        
+        BootCommand command = bootCommands.getCommands().get(0);
+        assertThat(command.getArguments().length, is(9));
+        assertEquals(command.getArguments()[0], "--restype");
+        assertEquals(command.getArguments()[1], "java.lang.String");
+        assertEquals(command.getArguments()[2], "-s");
+        assertEquals(command.getArguments()[3], "v");
+        assertEquals(command.getArguments()[4], "--name='custom-res'");
+        assertEquals(command.getArguments()[5], "--description=\"results \\\"in\\\" error\"");
+        assertEquals(command.getArguments()[6], "--property");
+        assertEquals(command.getArguments()[7], "value=\"${ENV=ini_ws_uri}\"");
+        assertEquals(command.getArguments()[8], "vfp/vfp-menu/ini.ws.uri");
     }
 
 }
